{{cookiecutter.project_slug}}/requirements.txt
-i https://pypi.org/simple
amqp==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
asgiref==3.2.7
attrs==19.3.0
automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
django-appconf==1.0.4
django-compressor==2.4
django-debug-toolbar==2.2
django-environ==0.4.5
django-libsass==0.8
django==3.0.5
flower==0.9.4
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
humanize==0.5.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
libsass==0.19.4
lxml==4.5.0
more-itertools==8.3.0; python_version >= '3.5'
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parsel==1.5.2
pillow==7.1.2
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protego==0.1.16
psycopg2==2.8.5
py==1.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydispatcher==2.0.5
pyhamcrest==2.0.2
pyopenssl==19.1.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-django==3.9.0
pytest==5.4.3
pytils==0.3
pytz==2020.1
pyyaml==5.3.1
queuelib==1.5.0
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
scrapy==2.1.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4; python_full_version >= '3.5.2'
twisted==20.3.0
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
w3lib==1.21.0
wcwidth==0.2.4
zope.interface==5.1.0
